Day 18 - Turtle And GUI/requirements.txt
colorgram.py==1.2.0
Pillow==8.1.0
